libbpf-rs: Introduce tracepoint module
