+++ libbpf-rs/src/lib.rs
@@ -93,6 +93,7 @@ pub mod query;
 mod ringbuf;
 mod skeleton;
 mod tc;
+mod tracepoint;
 mod user_ringbuf;
 mod util;
 mod xdp;
@@ -153,8 +154,6 @@ pub use crate::program::ProgramAttachType;
 pub use crate::program::ProgramImpl;
 pub use crate::program::ProgramMut;
 pub use crate::program::ProgramType;
-pub use crate::program::RawTracepointOpts;
-pub use crate::program::TracepointOpts;
 pub use crate::program::UprobeOpts;
 pub use crate::program::UsdtOpts;
 pub use crate::ringbuf::RingBuffer;
@@ -169,6 +168,8 @@ pub use crate::tc::TC_H_INGRESS;
 pub use crate::tc::TC_H_MIN_EGRESS;
 pub use crate::tc::TC_H_MIN_INGRESS;
 pub use crate::tc::TC_INGRESS;
+pub use crate::tracepoint::RawTracepointOpts;
+pub use crate::tracepoint::TracepointOpts;
 pub use crate::user_ringbuf::UserRingBuffer;
 pub use crate::user_ringbuf::UserRingBufferSample;
 pub use crate::util::num_possible_cpus;
