+++ libbpf-rs/src/program.rs
@@ -34,7 +34,9 @@ use crate::Error;
 use crate::ErrorExt as _;
 use crate::Link;
 use crate::Mut;
+use crate::RawTracepointOpts;
 use crate::Result;
+use crate::TracepointOpts;
 
 /// Options to optionally be provided when attaching to a uprobe.
 #[derive(Clone, Debug, Default)]
@@ -82,58 +84,6 @@ impl From<UsdtOpts> for libbpf_sys::bpf_usdt_opts {
             sz: size_of::<Self>() as _,
             usdt_cookie: cookie,
             // bpf_usdt_opts might have padding fields on some platform
-            ..Default::default()
-        }
-    }
-}
-
-/// Options to optionally be provided when attaching to a tracepoint.
-#[derive(Clone, Debug, Default)]
-pub struct TracepointOpts {
-    /// Custom user-provided value accessible through `bpf_get_attach_cookie`.
-    pub cookie: u64,
-    #[doc(hidden)]
-    pub _non_exhaustive: (),
-}
-
-impl From<TracepointOpts> for libbpf_sys::bpf_tracepoint_opts {
-    fn from(opts: TracepointOpts) -> Self {
-        let TracepointOpts {
-            cookie,
-            _non_exhaustive,
-        } = opts;
-
-        #[allow(clippy::needless_update)]
-        libbpf_sys::bpf_tracepoint_opts {
-            sz: size_of::<Self>() as _,
-            bpf_cookie: cookie,
-            // bpf_tracepoint_opts might have padding fields on some platform
-            ..Default::default()
-        }
-    }
-}
-
-/// Options to optionally be provided when attaching to a raw tracepoint.
-#[derive(Clone, Debug, Default)]
-pub struct RawTracepointOpts {
-    /// Custom user-provided value accessible through `bpf_get_attach_cookie`.
-    pub cookie: u64,
-    #[doc(hidden)]
-    pub _non_exhaustive: (),
-}
-
-impl From<RawTracepointOpts> for libbpf_sys::bpf_raw_tracepoint_opts {
-    fn from(opts: RawTracepointOpts) -> Self {
-        let RawTracepointOpts {
-            cookie,
-            _non_exhaustive,
-        } = opts;
-
-        #[allow(clippy::needless_update)]
-        libbpf_sys::bpf_raw_tracepoint_opts {
-            sz: size_of::<Self>() as _,
-            cookie,
-            // bpf_raw_tracepoint_opts might have padding fields on some platform
             ..Default::default()
         }
     }
